Remove redundant type casting

+++ neurodsp/spectral/power.py
@@ -214,7 +214,7 @@ def compute_spectrum_medfilt(sig, fs, filt_len=1., f_range=None):
     freqs = np.fft.fftfreq(len(sig), 1. / fs)[:int(np.ceil(len(sig) / 2.))]
 
     # Convert median filter length from Hz to samples, and make sure it is odd
-    filt_len_samp = int(int(filt_len / (freqs[1] - freqs[0])))
+    filt_len_samp = int(filt_len / (freqs[1] - freqs[0]))
     if filt_len_samp % 2 == 0:
         filt_len_samp += 1
